# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Note
  version: 0.0.1-SNAPSHOT
tags:
  - name: Note
    description: Заметки
paths:
  /note/create:
    post:
      tags:
        - CreateNote
      operationId: CreateNote
      description: Создание новой заметки
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteDTO"
      responses:
        '200':
          description: Note create successfully
  /note/update/{note_id}:
    put:
      tags:
        - UpdateNote
      operationId: UpdateNote
      description: Реадактирование заметки
      parameters:
        - name: note_id
          in: path
          description: Идентификатор заметки
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteDTO"
      responses:
        200:
          description: Note update successfully
  /note/delete/{note_id}:
    delete:
      tags:
        - DeleteNote
      operationId: DeleteNote
      description: Удаление заметки
      parameters:
        - name: note_id
          in: path
          description: Идентификатор заметки
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Note delete successfully
  /note/get/{note_id}:
    get:
      tags:
        - GetNote
      operationId: GetNote
      description: Получение заметки
      parameters:
        - name: note_id
          in: path
          description: Идентификатор заметки
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Note get successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoteDTO"
  /note/get/list/:
    get:
      tags:
        - GetNotes
      operationId: GetNotes
      description: Получение заметок

      responses:
        200:
          description: Notes get successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NoteDTO"

components:
  schemas:
    NoteDTO:
      title: NoteDTO
      type: object
      description: Модель заметки DTO
      required:
        - noteId
        - title
        - content
        - styleText
      properties:
        noteId:
          type: integer
          format: int64
          example: 1
          minimum: 1
          description: Номер заметки
        title:
          type: string
          example: Java internship
          description: Заголовок заметки
        content:
          type: string
          example: make a test task
          description: Текст заметки
        styleText:
          $ref: '#/components/schemas/StyleText'

    StyleText:
      title: StyleText
      type: object
      description: Модель стиля текста
      required:
        - italics
        - bold
        - font
        - textSize
      properties:
        italics:
          type: boolean
          example: false
          description: флаг курсива
        bold:
          type: boolean
          example: false
          description: флаг выделения жирным
        font:
          type: string
          example: Arial
          description: шрифт текста
        textSize:
          type: integer
          format: int32
          example: 12
          description: Размер шрифта